backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
asyncpg==0.30.0
redis==5.0.4
grpcio==1.80.0
protobuf==6.33.6
jinja2==3.1.4